Completed workouts, Dashboard, started routes
Finished the "core" swim, bike, run workouts. They are functional at
this point but do need to be visually tidied up. Also completed the
basic layout for the dashboard view. Will need to add the planned
workout information to the daily view and to the volume chart.
Began work on the routes with Google Maps