Added planned workouts and general workout views
Added the ability to update the view on the dashboard with the actual or
planned workouts for 7days via javascript and a select box. Also added a
Monthly Calendar view to Workouts. Now the user can navigate to any month
and see what days have workouts, edit/delete any of them and see a monthly
totals for swim, bike, run, strength and other on the views.