Add ability to add races and race reports
Added the ability for users to create races globally, then add a race to the
users's race calendar. Each user can then create a race report giving details
about the specific race. Also included personal bests that are automatically
taken from the database for each race type on the dashboard. Added an upcoming
races widget to the dashboard as well.